Add Imagen entity and remove legacy tpo-almacen
