+++ src/main/java/com/uade/tpo/almacen/entity/Imagen.java
@@ -1,0 +1,21 @@
+package com.uade.tpo.almacen.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Data
+@Entity
+public class Imagen {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column
+    private String imagen;
+
+    @ManyToOne
+    @JoinColumn(name = "producto_id", nullable = false)
+    private Producto producto;
+}
+
