+++ src/main/java/com/uade/tpo/almacen/excepciones/CategoriaNoEncontradaException.java
