+++ src/main/java/com/uade/tpo/almacen/excepciones/EstadoInvalidoException.java
