+++ src/main/java/com/uade/tpo/almacen/excepciones/NoEncontradoException.java
