+++ src/main/java/com/uade/tpo/almacen/excepciones/OperacionNoPermitidaException.java
