+++ src/main/java/com/uade/tpo/almacen/excepciones/UsuarioNoEncontradoException.java
