+++ src/main/java/com/uade/tpo/almacen/repository/CategoriaRepository.java
